V3.1.5  提交屏幕UI文件，修改电量显示逻辑，超声频率基础值由551K调整至580K

+++ Middlewares/LVGL/GUI_APP/lv_mainstart.h
@@ -44,7 +44,7 @@ extern uint8_t UltraDuty;
 extern uint8_t SlaveFlg;
 extern uint16_t FreqOffset;
 extern uint16_t LightLevel;
-extern uint32_t MotorLevel;
+extern uint16_t MotorLevel;
 extern SCREENID Screen_Id;
 extern SCREENSTATE ScreenState;
 extern DEVICESPARAM DevicesParam;
